+++ lightly_studio_view/src/lib/hooks/index.ts
@@ -42,3 +42,4 @@ export { useSubmitCombinationSelection } from '$lib/hooks/useSubmitCombinationSe
 export { useOperatorsDialog } from '$lib/hooks/useOperatorsDialog/useOperatorsDialog';
 export { useDeleteAnnotation } from '$lib/hooks/useDeleteAnnotation/useDeleteAnnotation';
 export { useSettings } from '$lib/hooks/useSettings';
+export { useAnnotationClassVisibility } from '$lib/hooks/useAnnotationClassVisibility/useAnnotationClassVisibility';