+++ lightly_studio_view/src/lib/hooks/useVideoFrameAnnotations/useVideoFrameAnnotations.ts
@@ -4,7 +4,7 @@ import calculateBinaryMaskFromRLE from '$lib/components/SampleAnnotation/SampleA
 
 interface FrameAnnotationDataURL {
     frameId: string;
-    annotationId: string;
+    sample_id: string;
     dataUrl: string;
     width: number;
     height: number;
@@ -67,7 +67,7 @@ export function useVideoFrameAnnotations({
 
                     frameAnnotations.push({
                         frameId: frame.sample_id,
-                        annotationId: segmentationAnnotation.sample_id,
+                        sample_id: segmentationAnnotation.sample_id,
                         dataUrl,
                         width: imageWidth,
                         height